re-organised the Flutter code. split upload.dart into 3 separate files. created a file for Message, Image and Video bubble respectively.
