+++ Flutter/lib/pages/camera_preview_screen.dart
@@ -1,20 +1,17 @@
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
-import 'dart:io';
-import 'package:path/path.dart' as path;
-import 'package:path_provider/path_provider.dart';
+import 'dart:developer' as developer;
 
 class CameraPreviewScreen extends StatefulWidget {
   final CameraController controller;
 
-  const CameraPreviewScreen({Key? key, required this.controller})
-      : super(key: key);
+  const CameraPreviewScreen({super.key, required this.controller});
 
   @override
-  _CameraPreviewScreenState createState() => _CameraPreviewScreenState();
+  CameraPreviewScreenState createState() => CameraPreviewScreenState();
 }
 
-class _CameraPreviewScreenState extends State<CameraPreviewScreen> {
+class CameraPreviewScreenState extends State<CameraPreviewScreen> {
   late CameraController _controller;
   late Future<void> _initializeControllerFuture;
 
@@ -35,35 +32,31 @@ class _CameraPreviewScreenState extends State<CameraPreviewScreen> {
     try {
       await _initializeControllerFuture;
 
-      final Directory appDir = await getApplicationDocumentsDirectory();
-
-
       XFile picture = await _controller.takePicture();
-
 
       Navigator.pop(context, picture);
     } catch (e) {
-      print(e);
+      developer.log(e.toString());
     }
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Take a Picture')),
+      appBar: AppBar(title: const Text('Take a Picture')),
       body: FutureBuilder<void>(
         future: _initializeControllerFuture,
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.done) {
             return CameraPreview(_controller);
           } else {
-            return Center(child: CircularProgressIndicator());
+            return const Center(child: CircularProgressIndicator());
           }
         },
       ),
       floatingActionButton: FloatingActionButton(
-        child: Icon(Icons.camera_alt),
         onPressed: _takePicture,
+        child: const Icon(Icons.camera_alt),
       ),
     );
   }
