+++ Flutter/lib/widgets/message_bubble.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+class MessageBubble extends StatelessWidget {
+  final String message;
+  final bool isMe;
+
+  const MessageBubble({super.key, required this.message, required this.isMe});
+
+  @override
+  Widget build(BuildContext context) {
+    final messageWidth = MediaQuery.of(context).size.width * 0.7;
+    final alignment = isMe ? MainAxisAlignment.end : MainAxisAlignment.start;
+
+    return Row(
+      mainAxisAlignment: alignment,
+      children: [
+        Container(
+          constraints: BoxConstraints(maxWidth: messageWidth),
+          decoration: BoxDecoration(
+            color: isMe
+                ? const Color(0xff373E4E)
+                : const Color.fromARGB(255, 58, 70, 82),
+            borderRadius: BorderRadius.circular(12),
+          ),
+          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+          margin: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
+          child: Text(
+            message,
+            style: const TextStyle(color: Colors.white),
+          ),
+        ),
+      ],
+    );
+  }
+}
